AuthService/requirements.txt
grpcio==1.56.0
grpcio-tools==1.56.0
protobuf==4.23.3
pyjwt==2.8.0
peewee==3.16.2
pymysql==1.0.3